add.c

+++ add.c
@@ -3,10 +3,10 @@
 #include<stdio.h>
 int main()
 {
-  int n1,n2,add;
+  float n1,n2,add;
   printf("Enter two numbers:");
-  scanf("%d%d",&n1,&n2);
+  scanf("%f%f",&n1,&n2);
   add=n1+n2;
-  printf("Addition of two numbers=%d",add);
+  printf("Addition of two numbers=%f",add);
   return 0;
 }
